backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-mall-backend"
version = "0.1.0"
description = "Smart Mall Management System — FastAPI backend"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [
	{ name = "SmartMall", email = "dev@smartmall.local" }
]

dependencies = [
	"fastapi>=0.115.0,<1.0.0",
	"uvicorn[standard]>=0.30.0,<1.0.0",
	"pydantic>=2.8.0,<3.0.0",
	"pydantic-settings>=2.4.0,<3.0.0",
	"python-dotenv>=1.0.0,<2.0.0",
	"python-jose[cryptography]>=3.3.0,<4.0.0",
	"python-multipart>=0.0.9,<1.0.0",
	"bcrypt>=4.1.0,<5.0.0",
	"email-validator>=2.0.0,<3.0.0",
	"motor>=3.6.0,<4.0.0",
	"beanie>=1.26.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.0.0,<9.0.0",
	"pytest-cov>=5.0.0,<6.0.0",
	"httpx>=0.27.0,<1.0.0",
	"ruff>=0.5.0,<1.0.0",
]

[project.urls]
Homepage = "https://example.com/smart-mall"
Repository = "https://example.com/smart-mall/repo"

# ─── Ruff ────────────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
	"E",   # pycodestyle errors
	"W",   # pycodestyle warnings
	"F",   # pyflakes
	"I",   # isort
	"B",   # flake8-bugbear
	"S",   # flake8-bandit
	"UP",  # pyupgrade
]
ignore = [
	"S105", # hardcoded-password-string (covered by env-based config)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ─── Pytest ──────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

# ─── Coverage ────────────────────────────────────────────────
[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true

